refactor: expose tutor Ask via clients.Gateway instead of aitutor package
- clients/gateway.go: Gateway gains an Anthropic client field and an
  Ask(question, chapterContext) method (moved from aitutor.Tutor);
  the prompt.md go:embed moves here too, since aitutor now depends
  on clients (for Gateway) and the reverse dependency would cycle.
- aitutor/cli.go: uses a package-level clients.Gateway (mirrors the
  chapter1/bot df.Bot package-var pattern) instead of a local Tutor.
- serve/web/web.go: askHandler goes through the same package-level
  Gateway instead of calling aitutor.Ask directly.
- clients/gateway_test.go: success, chapter-context formatting, and
  error-propagation tests using mocks/anthropic (moved from
  aitutor/tutor_test.go — mocks stay under the shared mocks/ tree,
  unchanged).
