+++ clients/prompt.md
